trRouting: Move the types defined in `TrRoutingService` to their own files
This will allow to import those types without necessarily importing the
file containing the calculation functions.
