+++ packages/chaire-lib-common/src/services/trRouting/TrRoutingResultConversion.ts
@@ -4,7 +4,7 @@
  * This file is licensed under the MIT License.
  * License text available at https://opensource.org/licenses/MIT
  */
-import { TrRoutingRoute } from './TrRoutingService';
+import { TrRoutingRoute } from './types';
 import { secondsSinceMidnightToTimeStr, secondsToMinutes } from '../../utils/DateTimeUtils';
 import { TrRoutingV2 } from '../../api/TrRouting';
